+++ CHANGELOG.md
@@ -17,6 +17,8 @@ _new implementation of fresa_
 #### [0.4.3] name (_1 jul 22_)
 
 - **added** - engine configuration variables using constexpr functions
+- **removed** - deprecated command line options in favor of the configuration file for now
+- **changed** - now unit tests are defined in the configuration file
 
 #### [0.4.2] numbers and letters (_29 jun 22_)